Initial Commit
you can normally save text data from individual files through the machine software. so i just dumped the raw values and matched the known values with the official text output. now with this commandline tool I could export all the pr1 files to a csv in batch, i needed the raw data because we were migrating to a manufacture and I wanted to maintain a record of the old data.